cleaning and commenting code
