+++ callbacks_preprocessing.py
@@ -1,5 +1,4 @@
 import dash
-import dash_bootstrap_components as dbc
 import json
 import utils
 import numpy as np
@@ -8,12 +7,13 @@ from app import variables
 from dash import Input, Output, State, callback, MATCH, ALL, html, ctx, dcc
 from definitions import ProcessTypology, EcgRemovalMethods, EnvelopeMethod
 from resurfemg import helper_functions as hf
-from scipy.signal import find_peaks
+
 
 card_counter = 0
 json_parameters = []
 
 
+# on loading add the emg graphs
 @callback(Output('preprocessing-original-container', 'children'),
           Input('load-preprocessing-div', 'data'))
 def show_raw_data(data):
@@ -30,6 +30,7 @@ def show_raw_data(data):
     return children_emg
 
 
+# apply the processing on the button click
 @callback(Output('preprocessing-processed-container', 'children'),
           Output('download-data-btn', 'disabled'),
           Input('apply-pipeline-btn', 'n_clicks'),
@@ -63,32 +64,38 @@ def show_data(click,
               additional_rem,
               additional_rem_idx):
 
-    save_data_enabled = False
-
+    # variables initialization
     global json_parameters
     json_parameters.clear()
 
     emg_data = variables.get_emg()
     sample_rate = variables.get_emg_freq()
 
+    # if data have been loaded, apply the processing
     if emg_data is not None:
+        # apply cut
         emg_cut = hf.bad_end_cutter_for_samples(emg_data, cut_percent, cut_tolerance)
         json_parameters.append(utils.build_cutter_params_json(1, cut_percent, cut_tolerance))
 
+        # apply filter
         emg_data_filtered = hf.emg_bandpass_butter_sample(emg_cut,
                                                           low_freq,
                                                           high_freq,
                                                           sample_rate)
         json_parameters.append(utils.build_bandpass_params_json(2, low_freq, high_freq))
 
+        # remove filtering artifacts (this is done automatically, no params are
+        # displayed in the GUI
         emg_cut_final = hf.bad_end_cutter_for_samples(emg_data_filtered, 3, 5)
         json_parameters.append(utils.build_cutter_params_json(3, 3, 5))
 
-        emg_ecg, titles = apply_ecg_removal(ecg_method, emg_cut_final, sample_rate)
+        # remove ECG
+        emg_ecg, titles = utils.apply_ecg_removal(ecg_method, emg_cut_final, sample_rate)
         json_parameters.append(utils.build_ecgfilt_params_json(4, EcgRemovalMethods(ecg_method)))
 
         new_step_emg = emg_ecg
 
+        # get the custom steps added, and apply the selected processing
         for n, card in enumerate(additional_card):
             card_id = card['index']
             step = additional_steps[n]
@@ -99,15 +106,22 @@ def show_data(click,
                 low_cut = additional_low[idx_low]
                 high_cut = additional_high[idx_high]
 
-                new_step_emg = hf.emg_bandpass_butter_sample(new_step_emg, low_cut, high_cut, sample_rate)
-                json_parameters.append(utils.build_bandpass_params_json(n+5, low_cut, high_cut))
+                new_step_emg = hf.emg_bandpass_butter_sample(new_step_emg,
+                                                             low_cut, high_cut,
+                                                             sample_rate)
+                json_parameters.append(utils.build_bandpass_params_json(len(json_parameters)+1,
+                                                                        low_cut,
+                                                                        high_cut))
 
             elif step == ProcessTypology.HIGH_PASS.value:
                 idx = utils.get_idx_dict_list(additional_low_idx, 'index', card_id)
                 low_cut = additional_low[idx]
 
-                new_step_emg = hf.emg_highpass_butter(new_step_emg, low_cut, sample_rate)
-                json_parameters.append(utils.build_highpass_params_json(n + 5, low_cut))
+                new_step_emg = hf.emg_highpass_butter(new_step_emg,
+                                                      low_cut,
+                                                      sample_rate)
+                json_parameters.append(utils.build_highpass_params_json(len(json_parameters)+1,
+                                                                        low_cut))
 
             elif step == ProcessTypology.LOW_PASS.value:
                 idx = utils.get_idx_dict_list(additional_high_idx, 'index', card_id)
@@ -122,7 +136,7 @@ def show_data(click,
 
                 ecg_additional_method = additional_rem[idx]
 
-                # at the moment we need to create a metrix with 3 leads to use the methods
+                # at the moment we need to create a matrix with 3 leads to use the methods
                 # the lead 0 is the  ecg lead, the other two are the same processed signal
                 # if the matrix is still bi-dimensional, we use it
 
@@ -131,29 +145,40 @@ def show_data(click,
                 else:
                     tmp_matrix = new_step_emg
 
-                new_step_emg, titles = apply_ecg_removal(ecg_additional_method, tmp_matrix, sample_rate)
-                json_parameters.append(utils.build_ecgfilt_params_json(n + 5, EcgRemovalMethods(ecg_additional_method)))
+                new_step_emg, titles = utils.apply_ecg_removal(ecg_additional_method,
+                                                         tmp_matrix,
+                                                         sample_rate)
+                json_parameters.append(utils.build_ecgfilt_params_json(len(json_parameters)+1,
+                                                                       EcgRemovalMethods(ecg_additional_method)
+                                                                       ))
 
-        emg_env = get_envelope(envelope_method, new_step_emg, sample_rate)
-        json_parameters.append(utils.build_envelope_params_json(len(json_parameters)+1, EnvelopeMethod(envelope_method)))
+        # At the end, extract the envelope
+        emg_env = utils.get_envelope(envelope_method, new_step_emg, sample_rate)
+        json_parameters.append(utils.build_envelope_params_json(len(json_parameters)+1,
+                                                                EnvelopeMethod(envelope_method)
+                                                                ))
 
+        # store the processed signal
         variables.set_emg_processed(emg_env)
-
+        # update the graphs
         children_emg = utils.add_emg_graphs(emg_env, sample_rate, titles)
-
+        # enable the data download
         save_data_enabled = False
-    else:
+    else: # if no data have been uploaded
         children_emg = []
         save_data_enabled = True
+
     return children_emg, save_data_enabled
 
 
+# open/close pipeline card
 @callback(Output('pipeline-card-body', 'is_open'),
           Input('pipeline-switch', 'value'))
 def show_raw_data(toggle_value):
     return toggle_value
 
 
+# open/close EMG graphs
 @callback(
     Output({"type": "emg-graph-collapse", "index": MATCH}, "is_open"),
     Input({"type": "emg-graph-switch", "index": MATCH}, "value"),
@@ -163,6 +188,7 @@ def collapse_graph(toggle_value):
     return toggle_value
 
 
+# add/remove custom steps
 @callback(Output('custom-preprocessing-steps', 'children'),
           Input('add-steps-btn', 'n_clicks'),
           Input({"type": "step-close-button", "index": ALL}, "n_clicks"),
@@ -176,28 +202,29 @@ def add_step(click, close, previous_content):
     if id_ctx is None:
         card_counter = 0
         return []
-
+    # if the add steps button is clicked add the card
     elif id_ctx == 'add-steps-btn':
         card_counter += 1
-        new_card = new_step_body(card_counter)
+        new_card = utils.get_new_step_body(card_counter)
 
         if previous_content is None:
             updated_content = new_card
         else:
             updated_content = previous_content + [new_card, html.P()]
-
+    # if the remove button is clicked, remove the card
     else:
         remove_idx = id_ctx['index']
         for n, el in enumerate(previous_content):
             if el['type'] == 'Card' and el['props']['id']['index'] == remove_idx:
-                del previous_content[n + 1]
-                previous_content.remove(el)
+                del previous_content[n + 1] # remove the html.P element
+                previous_content.remove(el) # remove the card
 
         updated_content = previous_content
 
     return updated_content
 
 
+# populate the options on the base of the selected processing type
 @callback(Output({"type": "additional-step-core", "index": MATCH}, "children"),
           Input({"type": "additional-step-type", "index": MATCH}, "value"),
           State({"type": "additional-step-core", "index": MATCH}, "id"),
@@ -218,88 +245,19 @@ def get_body(selected_value, card_id):
     return new_section
 
 
+# download the json file with the processing params and the processed emg
 @callback(Output('download-params', 'data'),
           Output('download-emg-processed', 'data'),
           Input('download-data-btn', 'n_clicks'),
           prevent_initial_call=True)
 def download_data(click):
 
+    # build the params file
     params_file = dict(content=json.dumps(json_parameters), filename='parameters.txt')
 
+    # build the csv file with the processed signal
+    # to use the dcc.Download element, we need to convert the np array into a dataframe
     df = pd.DataFrame(variables.get_emg_processed().transpose())
     emg_file = dcc.send_data_frame(df.to_csv, 'emg.csv')
 
     return params_file, emg_file
-
-
-def apply_ecg_removal(removal_method: int, emg_signal, sample_rate):
-    if removal_method == EcgRemovalMethods.ICA.value:
-        emg_ica = hf.compute_ICA_two_comp(emg_signal)
-        ecg_lead = emg_signal[0]
-
-        emg_ecg = hf.pick_lowest_correlation_array(emg_ica, ecg_lead)
-
-        titles = ["Filtered Track 2"]
-    elif removal_method == EcgRemovalMethods.GATING.value:
-        # TODO: change with QRS identification when available in library
-        peak_width = 0.001
-        peak_fraction = 0.40
-        ecg_rms = hf.full_rolling_rms(emg_signal[0, :], 10)
-        peak_height = peak_fraction * (max(ecg_rms) - min(ecg_rms))
-        ecg_peaks, _ = find_peaks(ecg_rms, height=peak_height, width=peak_width * sample_rate)
-
-        emg_ecg = hf.gating(emg_signal[2, :], ecg_peaks, method=0)
-
-        titles = ["Filtered Track 2"]
-    else:
-        emg_ecg = emg_signal
-        titles = None
-
-    return emg_ecg, titles
-
-
-def get_envelope(envelope_method: int, emg_signal, sample_rate):
-    if envelope_method == EnvelopeMethod.RMS.value:
-        # I set the window here to 100ms, but this may be changed
-        if emg_signal.ndim == 1:
-            emg_env = hf.full_rolling_rms(abs(emg_signal), int(sample_rate / 10))
-        else:
-            emg_env = np.array([hf.full_rolling_rms(lead, int(sample_rate / 10)) for lead in abs(emg_signal)])
-    elif envelope_method == EnvelopeMethod.FILTERING.value:
-        # THIS SHOULD BE CHANGED TO LOW PASS!
-        emg_env = hf.emg_highpass_butter(abs(emg_signal), 150, sample_rate)
-    else:
-        emg_env = emg_signal
-
-    return emg_env
-
-
-def new_step_body(index):
-    new_card = dbc.Card([
-        dbc.CardHeader([
-            html.Button(
-                html.I(className="fas fa-times", style={'color': 'red'}),
-                className="ml-auto close",
-                id={"type": "step-close-button", "index": str(index)},
-                style={'border': 'none',
-                       'background': 'transparent'}
-            ),
-            dbc.Label("Additional step")
-        ]),
-        dbc.Label("Step type"),
-        dbc.Select(
-            id={"type": "additional-step-type", "index": str(index)},
-            options=[
-                {"label": "", "value": "0"},
-                {"label": "Band-pass filter", "value": ProcessTypology.BAND_PASS.value},
-                {"label": "High-pass filter", "value": ProcessTypology.HIGH_PASS.value},
-                {"label": "Low-pass filter", "value": ProcessTypology.LOW_PASS.value},
-                {"label": "ECG removal", "value": ProcessTypology.ECG_REMOVAL.value},
-            ],
-            value="0"
-        ),
-        html.Div([], id={"type": "additional-step-core", "index": str(card_counter)})
-    ],
-        id={"type": "additional-step-card", "index": str(card_counter)})
-
-    return new_card
